fix: dark mode announcement bar, links
On dark mode, the announcement bar's color made it blend in with the background, so it was changed to light green and white text to match the landing button colors. Links within text blocks were not notable b/c they were white (blended in with normal text), so this was made to be neon blue to allow links to stand out more (consistent with light mode). Also, not sure why but felt like link hovers were jumpy, like another cursor type happens for a split second...so tried to force 'auto' for hovers.