feat: add 'how to build an nft' index page
There are a number of NFT-focused pages, but there is not a single one that unifies everything together. Adding this helps with SEO as well as simplifies link sharing. Additionlly, some of the SEO keywords per page have been updated, and the 'sidebars' also now include this index page in the 'Quickstarts' section.