feat: update SQL spec with INSERT + GROUP BY; Solidity bool<>integer
Tableland now allows a GROUP BY / HAVING in INSERT + SELECT subqueries; the spec and docs are adjusted accordingly. On a similar front, the spec used to say that Solidity bool data types should be converted to a SQLite integer when writing SQL in contracts; this is actually untrue since SQLite aliases TRUE/FALSE to a 1/0, which actually helps with Solidity<>SQL type parity.